requirements.txt
aiogram==2.25.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
alembic==1.13.1
